Adds the possibility to set a category to put the reverse complement of it's codon into the dna template.
Using cat-edit {catId} --reverse-complement 1 sets a category to be "reverse complement" if its getting put into the DNA template. This does not change the codon representation itself. To deactivate, use --reverse-complement 0.

For example, a DNA template with ACTGA{de1}ACTGA will put in "TT" if the codon of an element is "AA" and if the category is set to be reverse complement.